+++ src/main/java/pl/telvarost/mojangfixstationapi/mixin/client/misc/MinecraftMixin.java
@@ -24,8 +24,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import pl.telvarost.mojangfixstationapi.KeyBindingListener;
-import pl.telvarost.mojangfixstationapi.ModHelper;
 import pl.telvarost.mojangfixstationapi.client.MojangFixStationApiClientMod;
 import pl.telvarost.mojangfixstationapi.mixinterface.ChatScreenAccessor;